+++ LLM_API/Interfaces/IChatHandler.cs
@@ -1,0 +1,8 @@
+﻿using ChatManagement.Model.Chat;
+
+namespace LLM_API.Interfaces;
+
+public interface IChatHandler
+{
+    Task<ChatOutput> ProcessChat(ChatInput value);
+}
