--- RAG_API/Interfaces/IChatHandler.cs
@@ -1,9 +1,0 @@
-﻿using ChatManagement.Model;
-using ChatManagement.Model.Chat;
-
-namespace RAG_API.Interfaces;
-
-public interface IChatHandler
-{
-    Task<ChatOutput> ProcessChat(ChatInput value);
-}
